Refactor Producto and Lote DTOs; update endpoint for product search, enhance LoteDto with quantity handling, and adjust Venta repository to use Integer ID type

+++ src/main/java/com/farmacia/repository/VentaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.farmacia.models.Venta;
 
 @Repository
-public interface VentaRepository extends JpaRepository<Venta, Long> {
+public interface VentaRepository extends JpaRepository<Venta, Integer> {
 
     List<Venta> findBySucursalId(Integer sucursalId);
